Move AggregateExpr trait into physical-expr crate
